Move HydrogenRequest to foundation
