+++ packages/hydrogen/src/foundation/session/session.ts
@@ -1,7 +1,7 @@
 import {Logger} from '../../utilities/log';
 import {wrapPromise} from '../../utilities/suspense';
 import type {HydrogenResponse} from '../../framework/HydrogenResponse.server';
-import type {HydrogenRequest} from '../../framework/HydrogenRequest.server';
+import type {HydrogenRequest} from '../HydrogenRequest/HydrogenRequest.server';
 
 export type SessionSyncApi = {
   get: () => Record<string, string>;
