+++ packages/hydrogen/src/types.ts
@@ -1,6 +1,6 @@
 import type {ServerResponse} from 'http';
 import type {Logger, LoggerConfig} from './utilities/log/log';
-import type {HydrogenRequest} from './framework/HydrogenRequest.server';
+import type {HydrogenRequest} from './foundation/HydrogenRequest/HydrogenRequest.server';
 import type {HydrogenResponse} from './framework/HydrogenResponse.server';
 import type {
   Metafield,
